split PackageManager into PackageImporter and PackageExporter
