+++ src/PackageImporter.ts
@@ -1,11 +1,8 @@
-import { WriteStream } from 'fs';
 import fsExtra from 'fs-extra';
 import path from 'path';
 import promisepipe from 'promisepipe';
-import { Readable } from 'stream';
 import { dir } from 'tmp-promise';
 import yauzlPromise from 'yauzl-promise';
-import yazl from 'yazl';
 
 import ContentManager from './ContentManager';
 import EditorConfig from './EditorConfig';
@@ -13,20 +10,12 @@ import H5pError from './helpers/H5pError';
 import ValidationError from './helpers/ValidationError';
 import LibraryManager from './LibraryManager';
 import PackageValidator from './PackageValidator';
-import {
-    ContentId,
-    ITranslationService,
-    IUser,
-    Permission,
-    IH5PJson
-} from './types';
-import DependencyGetter from './DependencyGetter';
-import Library from './Library';
+import { ContentId, ITranslationService, IUser } from './types';
 
 /**
  * Handles the installation of libraries and saving of content from a H5P package.
  */
-export default class PackageManager {
+export default class PackageImporter {
     /**
      * @param {LibraryManager} libraryManager
      * @param {TranslationService} translationService
@@ -109,115 +98,6 @@ export default class PackageManager {
     }
 
     /**
-     * Creates a .h5p-package for the specified content file and pipes it to the stream.
-     * Throws H5pErrors if something goes wrong. The contents of the stream should be disregarded then.
-     * @param contentId The contentId for which the package should be created.
-     * @param outputStream The stream that the package is written to (e.g. the response stream fo Express)
-     */
-    public async createPackage(
-        contentId: ContentId,
-        outputStream: WriteStream,
-        user: IUser
-    ): Promise<void> {
-        if (!(await this.contentManager.contentExists(contentId))) {
-            throw new H5pError(
-                `Content can't be downloaded as no content with id ${contentId} exists.`
-            );
-        }
-        if (
-            !(await this.contentManager.getUserPermissions(
-                contentId,
-                user
-            )).some(p => p === Permission.Download)
-        ) {
-            throw new H5pError(
-                `You do not have permission to download content with id ${contentId}`
-            );
-        }
-        let contentStream: Readable;
-        try {
-            const content = await this.contentManager.loadContent(
-                contentId,
-                user
-            );
-            contentStream = new Readable();
-            contentStream._read = () => {
-                return;
-            };
-            contentStream.push(JSON.stringify(content));
-            contentStream.push(null);
-        } catch (error) {
-            throw new H5pError(
-                `Content can't be downloaded as the content data is unreadable.`
-            );
-        }
-        let metadataStream: Readable;
-        let metadata: IH5PJson;
-        try {
-            metadata = await this.contentManager.loadH5PJson(contentId, user);
-            metadataStream = new Readable();
-            metadataStream._read = () => {
-                return;
-            };
-            metadataStream.push(JSON.stringify(metadata));
-            metadataStream.push(null);
-        } catch (error) {
-            throw new H5pError(
-                `Content can't be downloaded as the content metadata is unreadable.`
-            );
-        }
-
-        const contentFiles = await this.contentManager.getContentFiles(
-            contentId,
-            user
-        );
-        const zip = new yazl.ZipFile();
-        zip.outputStream.pipe(outputStream);
-        zip.addReadStream(metadataStream, 'h5p.json');
-        zip.addReadStream(contentStream, 'content/content.json');
-        for (const contentFile of contentFiles) {
-            const filePath = path.join('content', contentFile);
-            zip.addReadStream(
-                this.contentManager.getContentFileStream(
-                    contentId,
-                    filePath,
-                    user
-                ),
-                filePath
-            );
-        }
-
-        const dependencyGetter = new DependencyGetter(this.libraryManager);
-        const mainLibrary = metadata.preloadedDependencies.find(
-            l => l.machineName === metadata.mainLibrary
-        );
-        if (!mainLibrary) {
-            throw new H5pError(
-                `The main library ${metadata.mainLibrary} is not listed in the dependencies`
-            );
-        }
-        const dependencies = await dependencyGetter.getDependencies(
-            new Library(
-                mainLibrary.machineName,
-                mainLibrary.majorVersion,
-                mainLibrary.minorVersion
-            ),
-            { editor: true, preloaded: true }
-        );
-        for (const dependency of dependencies) {
-            const files = await this.libraryManager.listFiles(dependency);
-            for (const file of files) {
-                zip.addReadStream(
-                    this.libraryManager.getFileStream(dependency, file),
-                    `${dependency.getDirName()}/${file}`
-                );
-            }
-        }
-
-        zip.end();
-    }
-
-    /**
      * Installs all libraries from the package. Assumes that the user calling this has the permission to install libraries!
      * Throws errors if something goes wrong.
      * @param {string} packagePath The full path to the H5P package file on the local disk.
@@ -258,7 +138,7 @@ export default class PackageManager {
             await packageValidator.validatePackage(packagePath, false, true); // no need to check result as the validator throws an exception if there is an error
             const { path: tempDirPath } = await dir(); // we don't use withDir here, to have better error handling (catch block below)
             try {
-                await PackageManager.extractPackage(packagePath, tempDirPath, {
+                await PackageImporter.extractPackage(packagePath, tempDirPath, {
                     includeContent: copyContent,
                     includeLibraries: installLibraries,
                     includeMetadata: copyContent
@@ -287,7 +167,7 @@ export default class PackageManager {
                 if (copyContent) {
                     if (!this.contentManager) {
                         throw new Error(
-                            'PackageManager was initialized with a ContentManager, but you want to copy content from a package. Pass a ContentManager object to the the constructor!'
+                            'PackageImporter was initialized with a ContentManager, but you want to copy content from a package. Pass a ContentManager object to the the constructor!'
                         );
                     }
                     newContentId = await this.contentManager.copyContentFromDirectory(
@@ -315,4 +195,4 @@ export default class PackageManager {
     }
 }
 
-module.exports = PackageManager;
+module.exports = PackageImporter;
