+++ test/PackageImporter.test.ts
@@ -2,19 +2,18 @@ import * as fsExtra from 'fs-extra';
 import * as path from 'path';
 import promisepipe from 'promisepipe';
 import { BufferWritableMock } from 'stream-mock';
-import { withDir, withFile } from 'tmp-promise';
-import yauzlPromise from 'yauzl-promise';
+import { withDir } from 'tmp-promise';
 
 import ContentManager from '../src/ContentManager';
 import EditorConfig from '../src/EditorConfig';
 import FileContentStorage from '../src/FileContentStorage';
 import FileLibraryStorage from '../src/FileLibraryStorage';
 import LibraryManager from '../src/LibraryManager';
-import PackageManager from '../src/PackageManager';
+import PackageImporter from '../src/PackageImporter';
 import TranslationService from '../src/TranslationService';
 import User from '../src/User';
 
-describe('basic package manager functionality', () => {
+describe('package importer', () => {
     it('installs libraries', async () => {
         await withDir(
             async ({ path: tmpDirPath }) => {
@@ -24,12 +23,12 @@ describe('basic package manager functionality', () => {
                 const libraryManager = new LibraryManager(
                     new FileLibraryStorage(libraryDir)
                 );
-                const packageManager = new PackageManager(
+                const packageImporter = new PackageImporter(
                     libraryManager,
                     new TranslationService({}),
                     new EditorConfig(null)
                 );
-                await packageManager.installLibrariesFromPackage(
+                await packageImporter.installLibrariesFromPackage(
                     path.resolve('test/data/validator/valid2.h5p')
                 );
 
@@ -67,13 +66,13 @@ describe('basic package manager functionality', () => {
                 const libraryManager = new LibraryManager(
                     new FileLibraryStorage(libraryDir)
                 );
-                const packageManager = new PackageManager(
+                const packageImporter = new PackageImporter(
                     libraryManager,
                     new TranslationService({}),
                     new EditorConfig(null),
                     contentManager
                 );
-                const contentId = await packageManager.addPackageLibrariesAndContent(
+                const contentId = await packageImporter.addPackageLibrariesAndContent(
                     path.resolve('test/data/validator/valid2.h5p'),
                     user
                 );
@@ -112,77 +111,5 @@ describe('basic package manager functionality', () => {
             },
             { keep: false, unsafeCleanup: true }
         );
-    });
-    it('creates h5p packages', async () => {
-        await withDir(
-            async ({ path: tmpDirPath }) => {
-                const contentDir = path.join(tmpDirPath, 'content');
-                const libraryDir = path.join(tmpDirPath, 'libraries');
-                await fsExtra.ensureDir(contentDir);
-                await fsExtra.ensureDir(libraryDir);
-
-                const user = new User();
-                user.canUpdateAndInstallLibraries = true;
-
-                const contentManager = new ContentManager(
-                    new FileContentStorage(contentDir)
-                );
-                const libraryManager = new LibraryManager(
-                    new FileLibraryStorage(libraryDir)
-                );
-                const packageManager = new PackageManager(
-                    libraryManager,
-                    new TranslationService({}),
-                    new EditorConfig(null),
-                    contentManager
-                );
-                const contentId = await packageManager.addPackageLibrariesAndContent(
-                    path.resolve('test/data/validator/valid2.h5p'),
-                    user
-                );
-
-                await withFile(
-                    async fileResult => {
-                        const writeStream = fsExtra.createWriteStream(
-                            fileResult.path
-                        );
-                        await packageManager.createPackage(
-                            contentId,
-                            writeStream,
-                            user
-                        );
-                        await new Promise(async (resolve, reject) => {
-                            const whenStreamClosed = jest.fn();
-                            writeStream.on('close', whenStreamClosed);
-                            writeStream.on('close', async () => {
-                                expect(whenStreamClosed).toBeCalled();
-
-                                const oldZipFile = await yauzlPromise.open(
-                                    path.resolve(
-                                        'test/data/validator/valid2.h5p'
-                                    )
-                                );
-                                const oldEntries = (await oldZipFile.readEntries())
-                                    .map(e => e.fileName)
-                                    .sort();
-
-                                const newZipFile = await yauzlPromise.open(
-                                    fileResult.path
-                                );
-                                const newEntries = (await newZipFile.readEntries())
-                                    .map(e => e.fileName)
-                                    .sort();
-
-                                expect(newEntries).toMatchObject(oldEntries);
-
-                                resolve();
-                            });
-                        });
-                    },
-                    { postfix: '.h5p', keep: false }
-                );
-            },
-            { keep: false, unsafeCleanup: true }
-        );
-    });
+    });    
 });
